backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pinecone
google-genai
pypdf